ex00 finished

+++ cpp_09/ex00/BitcoinExchange.hpp
@@ -2,12 +2,13 @@
 #define BITCOIN_EXCHANGE_HPP
 
 #include <iostream>
+#include <iomanip>
 #include <algorithm>
 #include <map>
 #include <cmath>
 #include <ctime>
 #include <cstring>
- 
+
 #include <iostream>
 #include <string>
 #include <fstream>
@@ -20,7 +21,7 @@ class BitcoinExchange {
         BitcoinExchange const &operator=(BitcoinExchange const &);
 
         int     generateDatabase(void);
-        double  calcExchangeRate(struct tm date, double btc);
+        float  calcExchangeRate(struct tm date, double btc);
 
     private:
         std::map<std::string, float>	_db;
